requirements.txt
# 图像处理美颜与质量评价系统依赖包
# 版本要求：Python 3.6+

# 图形用户界面
PyQt5>=5.15.0

# 图像处理核心库
opencv-python>=4.5.0
numpy>=1.19.0
Pillow>=8.0.0

# 人脸状态检测依赖
mediapipe>=0.10.0

# 可选：用于更高级的图像处理（未来扩展）
# scikit-image>=0.18.0
# matplotlib>=3.3.0